+++ src/test/java/com/github/holiver98/service/inmemory/InMemoryCartServiceTestBase.java
@@ -5,6 +5,7 @@ import com.github.holiver98.model.IngredientType;
 import com.github.holiver98.model.Pizza;
 import com.github.holiver98.model.PizzaSize;
 
+import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -13,19 +14,19 @@ public class InMemoryCartServiceTestBase {
         Pizza pizza = new Pizza();
         pizza.setName(name);
         pizza.setSize(PizzaSize.NORMAL);
-        pizza.setPrice(20.5f);
+        pizza.setPrice(BigDecimal.valueOf(20.5));
         pizza.setCustom(false);
         pizza.setId(id);
 
         Set<Ingredient> ingredients = new HashSet<Ingredient>();
         Ingredient tomatoSauce = new Ingredient();
         tomatoSauce.setName("tomato sauce");
-        tomatoSauce.setPrice(2.1f);
+        tomatoSauce.setPrice(BigDecimal.valueOf(2.1));
         tomatoSauce.setType(IngredientType.PIZZA_BASESAUCE);
 
         Ingredient cheese = new Ingredient();
         cheese.setName("mozarella");
-        cheese.setPrice(3.3f);
+        cheese.setPrice(BigDecimal.valueOf(3.3));
         cheese.setType(IngredientType.PIZZA_TOPPING);
 
         ingredients.add(tomatoSauce);
@@ -39,23 +40,23 @@ public class InMemoryCartServiceTestBase {
         Pizza pizza = new Pizza();
         pizza.setName("twosaucepizza");
         pizza.setSize(PizzaSize.NORMAL);
-        pizza.setPrice(20.5f);
+        pizza.setPrice(BigDecimal.valueOf(20.5));
         pizza.setCustom(false);
 
         Set<Ingredient> ingredients = new HashSet<Ingredient>();
         Ingredient tomatoSauce = new Ingredient();
         tomatoSauce.setName("tomato sauce");
-        tomatoSauce.setPrice(2.1f);
+        tomatoSauce.setPrice(BigDecimal.valueOf(2.1));
         tomatoSauce.setType(IngredientType.PIZZA_BASESAUCE);
 
         Ingredient sourCream = new Ingredient();
         sourCream.setName("sour cream");
-        sourCream.setPrice(1.1f);
+        sourCream.setPrice(BigDecimal.valueOf(1.1));
         sourCream.setType(IngredientType.PIZZA_BASESAUCE);
 
         Ingredient cheese = new Ingredient();
         cheese.setName("mozarella");
-        cheese.setPrice(3.3f);
+        cheese.setPrice(BigDecimal.valueOf(3.3));
         cheese.setType(IngredientType.PIZZA_TOPPING);
 
         ingredients.add(tomatoSauce);
